--- src/App.css
@@ -1,24 +1,0 @@
-.App-logo {
-    animation: spin infinite 20s linear;
-    height: 80px;
-  }
-  
-  .App-header {
-    background-color: #222;
-    padding: 20px;
-    color: white;
-    text-align: center;
-  }
-  
-  .App-intro {
-    font-size: large;
-  }
-  
-  @keyframes spin {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
-  }
