+++ src/components/Order/Orders.js
@@ -5,7 +5,7 @@ import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
-import Title from './Title';
+import Title from '../../componentsUtils/Title.js';
 
 // Generate Order Data
 function createData(id, date, name, shipTo, paymentMethod, amount) {
